Windows: Fixed location of imported libraries for Ninja.
The Ninja generator does not emit binaries to "Debug" or "Release"
folders (unlike the Visual Studio generator), as it is not a
multi-configuration generator. To fix this, we explicitly set the
`IMPORTED_LOCATION` instead of `IMPORTED_LOCATION_(DEBUG|RELEASE)`
when using the Ninja generator on Windows. This does not reuse the
location used on non-Windows platforms, as they're not consistent
despite the generator being the same.

Review: https://reviews.apache.org/r/65720