Windows: Fixed remaining W* macros in `windows.hpp`.
The `WIFEXITED` and `WIFSIGNALED` were incorrectly checking if the
exit code was not -1 to determine if the process exited or was signaled.
However, -1 is a valid return code on Windows, so places in the Mesos
codebase that called `CHECK(WIFEXITED(status)|| WIFSIGNALED(status))`
would end up aborting the agent or executor.

Review: https://reviews.apache.org/r/65840