Windows: Replaced `_wopen` with `CreateFileW` in `os::open`.
Instead of using the CRT implementation of `_wopen` for the
`os::open()` API, we now use the Windows API `CreateFileW`, mapping
each of the Linux `open` flags to their semantic equivalents. This
will make implementing overlapped I/O possible, and is a step toward
removing the use of integer file descriptors on Windows.

Note that instead of redefining the C flags like `O_RDONLY`, we just
use them directly in our mapping logic, and set the used but
unsupported flags to zero.

This change uncovered several bugs such as incorrect access flags, and
used-but-not-included headers.

We currently ignore creation permissions as they will be handled in a
broader project to map permissions to Windows correctly.