Windows: Refactored `subprocess_windows.cpp` to use `os::open`.
Previously, `os::open` used the CRT function `_wopen`, and so this
file was written to use the `CreateFile` API directly. Now that
`os::open` uses the Windows API, all this duplicate code can be
deleted in favor of using the `os::open` and
`internal::windows::set_inherit`. The major benefit here is that the
logic now almost exactly matches the POSIX counterpart in
`subprocess_posix.cpp`, to the point that we may want to recombine
these files in the future.