Added `io::prepare_async` and `io::is_async` functions for libprocess.
`os::nonblock` and `os::isNonBlock` are stubs that only do things for
Windows sockets. For the IOCP implementation, we need to associate the
handles with the IOCP hande, which is created by libprocess. Thus, we
need these new functions to prepare the handle for the async IO
functions by either setting the hande to non-blocking on UNIX systems
or by associating it with an IOCP handle on Windows.