Windows: Wrapped docker executor in job object.
One of the docker tests revealed that the docker executor could leak
some of the docker processes if it gets killed before the docker
process completes. To avoid this, we run the docker executor in a job
object, so that all child processes of it die when the executor dies.